Beta version of Magic Map
Known issues: It's too big! But there appears to be no way to reduce font below 10, which is how the issue is solved on mudlet. On web client, the map scrolls based on character location, which is how I would like to solve the issue. Related to this is the flashing that currently occurs - limiting amount of map shown would likely solve this as well. It's on my todo list! But even as is, it's semi-usable.